Change decimals variable

+++ contracts/controller/DAOToken.sol
@@ -3,8 +3,6 @@ pragma solidity ^0.4.21;
 import "zeppelin-solidity/contracts/token/ERC20/MintableToken.sol";
 import "zeppelin-solidity/contracts/token/ERC20/BurnableToken.sol";
 import "zeppelin-solidity/contracts/token/ERC827/ERC827Token.sol";
-import "zeppelin-solidity/contracts/lifecycle/Destructible.sol";
-
 
 
 /**
@@ -16,7 +14,7 @@ contract DAOToken is ERC827Token,MintableToken,BurnableToken {
 
     string public name;
     string public symbol;
-    uint public constant DECIMAL = 18;
+    uint8 public constant decimal = 18;
     uint public cap;
 
     /**
